+++ langchain/src/agents/agent_toolkits/index.ts
@@ -1,7 +1,7 @@
-export { JsonToolkit, createJsonAgent } from "./json.js";
-export { SqlToolkit, createSqlAgent } from "./sql.js";
+export { JsonToolkit, createJsonAgent } from "./json/json.js";
+export { SqlToolkit, createSqlAgent } from "./sql/sql.js";
 export {
   RequestsToolkit,
   OpenApiToolkit,
   createOpenApiAgent,
-} from "./openapi.js";
+} from "./openapi/openapi.js";
