+++ langchain/src/agents/agent_toolkits/openapi/openapi.ts
@@ -1,22 +1,22 @@
-import { BaseLLM } from "../../llms/index.js";
+import { BaseLLM } from "../../../llms/index.js";
 import {
   DynamicTool,
   JsonSpec,
   RequestsGetTool,
   RequestsPostTool,
   Tool,
-} from "../tools/index.js";
-import { AgentExecutor } from "../executor.js";
+} from "../../tools/index.js";
+import { AgentExecutor } from "../../executor.js";
 import {
   OPENAPI_PREFIX,
   OPENAPI_SUFFIX,
   JSON_EXPLORER_DESCRIPTION,
 } from "./prompt.js";
-import { LLMChain } from "../../chains/index.js";
-import { CreatePromptArgs, ZeroShotAgent } from "../mrkl/index.js";
-import { Toolkit } from "./base.js";
-import { Headers } from "../tools/requests.js";
-import { createJsonAgent, JsonToolkit } from "./json.js";
+import { LLMChain } from "../../../chains/index.js";
+import { CreatePromptArgs, ZeroShotAgent } from "../../mrkl/index.js";
+import { Toolkit } from "../base.js";
+import { Headers } from "../../tools/requests.js";
+import { createJsonAgent, JsonToolkit } from "../json/json.js";
 
 export class RequestsToolkit extends Toolkit {
   tools: Tool[];
