+++ langchain/src/agents/agent_toolkits/sql/sql.ts
@@ -5,14 +5,14 @@ import {
   QueryCheckerTool,
   QuerySqlTool,
   SqlDatabase,
-} from "../tools/index.js";
-import { Toolkit } from "./base.js";
-import { BaseLLM } from "../../llms/index.js";
+} from "../../tools/index.js";
+import { Toolkit } from "../base.js";
+import { BaseLLM } from "../../../llms/index.js";
 import { SQL_PREFIX, SQL_SUFFIX } from "./prompt.js";
-import { interpolateFString } from "../../prompts/template.js";
-import { LLMChain } from "../../chains/index.js";
-import { ZeroShotAgent, CreatePromptArgs } from "../mrkl/index.js";
-import { AgentExecutor } from "../executor.js";
+import { interpolateFString } from "../../../prompts/template.js";
+import { LLMChain } from "../../../chains/index.js";
+import { ZeroShotAgent, CreatePromptArgs } from "../../mrkl/index.js";
+import { AgentExecutor } from "../../executor.js";
 
 type SqlCreatePromptArgs = {
   /** Number of results to return. */
